New podcasts corrections, plus jhove check condition, plus deleting empty sip folder condition

+++ scripts/podcasts.py
@@ -293,9 +293,10 @@ class Podcast_pipeline():
 						logger.error(str(e))
 					if os.path.isfile(os.path.join(rosetta_folder,epis["serial_mms"],"done")):
 						os.remove(os.path.join(rosetta_folder,epis["serial_mms"],"done"))
-					if os.listdir(os.path.join(rosetta_folder, epis["serial_mms"],"content","streams")) == []:
-						shutil.rmtree(os.path.join(rosetta_folder, epis["serial_mms"],"content","streams"))
-						logger.info("SIP folder removed - {}".format(os.path.join(rosetta_folder, epis["serial_mms"],"content","streams")))
+					if os.path.isdir(os.path.join(rosetta_folder, epis["serial_mms"],"content","streams")):
+						if os.listdir(os.path.join(rosetta_folder, epis["serial_mms"],"content","streams")) == []:
+							shutil.rmtree(os.path.join(rosetta_folder, epis["serial_mms"],"content","streams"))
+							logger.info("SIP folder removed - {}".format(os.path.join(rosetta_folder, epis["serial_mms"],"content","streams")))
 					
 
 
@@ -455,17 +456,17 @@ class Podcast_pipeline():
 		# self.get_ies_from_reports()
 		# lst = self.read_ies_file()
 		# self.insert_ies()
-		# self.finish_existing_records_and_delete_files("prod")
-		#self.db_handler.update_the_last_issue()
-		# self.db_handler.delete_done_from_db()
+		self.finish_existing_records_and_delete_files("prod")
+		self.db_handler.update_the_last_issue()
+		self.db_handler.delete_done_from_db()
 		self.update_database_from_spreadsheetand_delete_row()
-		# Set "sb" if whould like records in "sb"
-		# Set my_rec.record_creating_routune(update = True) to update records with existing mms id.
-		# Be careful not to update record with SB mms_id in Production. Normally SB is updating regularly by making Production copy.
-		# my_rec = RecordCreator(self.alma_key)
-		# my_rec.record_creating_routine()
-		# sip_routine()
-		# harvest()
+		### Set "sb" if whould like records in "sb"
+		### Set my_rec.record_creating_routune(update = True) to update records with existing mms id.
+		### Be careful not to update record with SB mms_id in Production. Normally SB is updating regularly by making Production copy.
+		my_rec = RecordCreator(self.alma_key)
+		my_rec.record_creating_routine()
+		sip_routine()
+		harvest()
 		 		
 def main():
